Revert workflow changes; move setting from ServiceBasedMerchant to Options; fix tests.
